Extract User concerns
Make new user concerns for `User.unused`, `User.limited_profile` and
`User.limited_profile?`.
