+++ app/models/user.rb
@@ -44,11 +44,13 @@ class User < ApplicationRecord
   include AlaveteliFeatures::Helpers
   include AlaveteliPro::PhaseCounts
   include User::Authentication
+  include User::LimitedProfile
   include User::LoginToken
   include User::OneTimePassword
   include User::Slug
   include User::SpreadableAlerts
   include User::Survey
+  include User::Unused
   include Taggable
   include Rails.application.routes.url_helpers
   include LinkToHelper
@@ -299,69 +301,6 @@ class User < ApplicationRecord
   def self.find_similar_named_users(user)
     User.where('name ILIKE ? AND email_confirmed = ? AND id <> ?',
                 user.name, true, user.id).order(:created_at)
-  end
-
-  # return an array of limited users i.e. untrusted users without any requests
-  # or classifications
-  def self.limited_profile
-    roles = Role.arel_table
-    user_roles = Arel::Table.new(:users_roles)
-    user_admin_roles_exists = user_roles.project(1).
-      join(roles).on(user_roles[:role_id].eq(roles[:id])).
-      where(user_roles[:user_id].eq(User.arel_table[:id])).
-      where(roles[:name].in(%w[admin pro_admin])).
-      exists
-
-    User.
-      where.not(email: internal_admin_user.email).
-      where(confirmed_not_spam: false).
-      where(info_requests_count: 0).
-      where(status_update_count: 0).
-      where.not(user_admin_roles_exists)
-  end
-
-  # return an array of unused users i.e. users without any user generated
-  # content, without granted an user role or recent sign ins
-  def self.unused
-    citations = Citation.arel_table
-    citations_exists = citations.project(1).
-      where(citations[:user_id].eq(User.arel_table[:id])).
-      exists
-
-    # don't return admins, pro admins, pros
-    roles = Role.arel_table
-    user_roles = Arel::Table.new(:users_roles)
-    user_roles_exists = user_roles.project(1).
-      join(roles).on(user_roles[:role_id].eq(roles[:id])).
-      where(user_roles[:user_id].eq(User.arel_table[:id])).
-      where(roles[:name].in(%w[admin pro_admin pro])).
-      exists
-
-    # don't return users who have previous submissions to a project
-    submission = Project::Submission.arel_table
-    submission_exists = submission.project(1).
-      where(submission[:user_id].eq(User.arel_table[:id])).
-      exists
-
-    # don't return users who have signed in recently
-    sign_ins = User::SignIn.arel_table
-    sign_ins_exists = sign_ins.project(1).
-      where(sign_ins[:user_id].eq(User.arel_table[:id])).
-      exists
-
-    User.
-      where.not(email: internal_admin_user.email).
-      where(info_requests_count: 0).
-      where(info_request_batches_count: 0).
-      where(request_classifications_count: 0).
-      where(status_update_count: 0).
-      where(track_things_count: 0).
-      where(comments_count: 0).
-      where(public_body_change_requests_count: 0).
-      where.not(citations_exists).
-      where.not(user_roles_exists).
-      where.not(submission_exists).
-      where.not(sign_ins_exists)
   end
 
   def view_hidden?
@@ -636,14 +575,6 @@ class User < ApplicationRecord
       self.profile_photo = new_profile_photo
       save!
     end
-  end
-
-  def limited_profile?
-    self != User.internal_admin_user &&
-      !is_admin? && !is_pro_admin? &&
-      !confirmed_not_spam? &&
-      info_requests_count.zero? &&
-      status_update_count.zero?
   end
 
   def show_profile_photo?
